+++ src/main/java/simulator/model/MapInfo.java
@@ -3,8 +3,8 @@ package simulator.model;
 public interface MapInfo extends JSONable {
     public int getCols();
     public int getRows();
-    public double getWidth();
-    public double getHeight();
+    public int getWidth();
+    public int getHeight();
     public double getRegionWidth();
     public double getRegionHeight();
 }
